Converted error reporting to exceptions. Closes #9.
So far we had used either assert statements or if/print/exit
throughout the code. This made testing for errors impossible. Now we
throw exceptions for all error cases.

There are a few assert statements remaining, but those are for
checking internal logic. For example, `eval_if` asserts that the list
passed to it is indeed an `if` form. This is just an extra precaution
since we've already seen it is an `if` form and sent it to `eval_if`.