Added function forms of some primitives to stdlib. Closes #22.
Added function forms of prep, first, rest, typeof and eq to
stdlib. These functions work exactly like the primitives and if fact
work by immediately calling the corresponding primitives. With these
in place, now we can for example send `first` as the first argument of
`mapf`.

`if` and `quote` equivalents were not added since neither can work as
functions.