Nil is now considered `eq` to nil.
I didn't originally planned for this; nil wasn't supposed to be a
special value. It was supposed to be an empty list and treated like
any other list. But since we are using nil as falsity, we need it to
be eq to itself or else it would be useless. Perhaps we need to
introduce special false and true values later.